refactor(analytics): moved analytics folder to the public level
